+++ common/meta/api/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+common-base = {path = "../../base" }
 common-datavalues = {path = "../../datavalues" }
 common-exception = {path = "../../exception" }
 common-meta-types = {path= "../types" }
